requirements.txt
tensorflow-cpu==2.15.0
fastapi==0.110.0
uvicorn==0.27.1
pillow==10.3.0
numpy==1.26.4
python-multipart==0.0.9
